+++ tests/descriptor/test_base.py
@@ -163,7 +163,7 @@ def test_base_descriptor_2(data):
 def test_base_descriptor_3(data):
     bd = data['descriptor']()
     try:
-        bd.fit([1, 2, 3, 4], )
+        bd.fit([1, 2, 3, 4], ),
     except TypeError:
         assert True
     else:
