+++ xenonpy/descriptor/base.py
@@ -330,7 +330,7 @@ class BaseDescriptor(BaseEstimator, TransformerMixin, metaclass=TimedMetaClass):
         X = self._check_input(X)
         for k, features in self.__featurizers__.items():
             for f in features:
-                f.fit(X[k], )
+                f.fit(X[k], ),
 
         return self
 
@@ -351,7 +351,7 @@ class BaseDescriptor(BaseEstimator, TransformerMixin, metaclass=TimedMetaClass):
         X = self._check_input(X)
         for k, features in self.__featurizers__.items():
             for f in features:
-                ret = f.transform(X[k])
+                ret = f.proposal(X[k])
                 if isinstance(ret, list):
                     ret = _make_df(f, ret)
                 results.append(ret)
