+++ xenonpy/inverse/__init__.py
@@ -2,7 +2,6 @@
 #  Use of this source code is governed by a BSD-style
 #  license that can be found in the LICENSE file.
 
-__all__ = ['BaseSMC', 'IQSPR', 'BaseProposer', 'BaseLogLikelihood']
+__all__ = ['IQSPR']
 
 from .iqspr import IQSPR
-from .smc import BaseSMC
