--- xenonpy/utils/__init__.py
@@ -1,10 +1,0 @@
-#  Copyright (c) 2019. yoshida-lab. All rights reserved.
-#  Use of this source code is governed by a BSD-style
-#  license that can be found in the LICENSE file.
-
-__all__ = [
-    'set_env', 'get_data_loc', 'get_conf', 'get_dataset_url', 'get_sha256', 'absolute_path', 'Switch',
-    'Timer', 'TimedMetaClass'
-]
-from .base_class import *
-from .gadget import *
